backend/requirements.txt
anthropic==0.94.0
fastapi==0.135.3
numpy==2.4.4
python-dotenv==1.2.2
rank-bm25==0.2.2
sentence-transformers==5.4.0
uvicorn[standard]==0.44.0
